Renamed files and added colours
