+++ frontend/next.config.js
@@ -1,7 +1,12 @@
+const path = require('path');
+
 /** @type {import('next').NextConfig} */
 const nextConfig = {
 	reactStrictMode: true,
 	output: 'standalone',
+	sassOptions: {
+		includePaths: [path.join(__dirname, 'src/styles')],
+	},
 	async rewrites() {
 		return {
 		  beforeFiles: [
@@ -14,7 +19,7 @@ const nextConfig = {
 			},
 		  ]
 		}
-	  } 
+	}
 };
 
 module.exports = nextConfig;
