+++ frontend/src/components/BookingCard/BookingCard.module.scss
@@ -1,5 +1,9 @@
+@import 'src/styles/variables';
+
 .card {
 	box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
 	padding: 1rem;
 	border-radius: 5px;
+	background-color: $black;
+
 }
