+++ frontend/src/components/GameCard/GameCard.module.scss
@@ -1,3 +1,5 @@
+@import 'src/styles/variables';
+
 .card {
 	box-shadow: 0 4px 10px rgba(0, 0, 0, 0.2);
 	padding: 1rem;
