+++ frontend/src/components/GameFilter/GameFilter.module.scss
