+++ frontend/src/components/GamesList/GamesList.module.scss
