+++ frontend/src/components/SiteAdminComponents/OrganizationList/OrganizationList.tsx
@@ -2,7 +2,7 @@ import { ChangeEvent, FC, useState } from 'react';
 import { useQuery, useQueryClient } from '@tanstack/react-query';
 import axios from 'axios';
 import { AxiosError } from 'axios';
-import styles from './OrganizationList.module.css';
+import styles from './OrganizationList.module.scss';
 import debounce from 'lodash.debounce';
 import OrganizationListEntry from '../OrganizationListEntry/OrganizationListEntry';
