+++ frontend/src/components/SuggestionCard/SuggestionCard.module.scss
