+++ frontend/src/components/SuggestionsList/SuggestionsList.module.scss
