+++ frontend/src/pages/_app.tsx
@@ -1,6 +1,6 @@
 import type { AppProps } from 'next/app';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import '../styles/globals.css';
+import '../styles/globals.scss';
 
 const queryClient = new QueryClient();
