+++ frontend/src/pages/index.module.scss
