+++ frontend/src/pages/suggestion.module.scss
