+++ frontend/src/styles/globals.scss
@@ -1,3 +1,5 @@
+@import 'variables';
+
 * {
 	box-sizing: border-box;
 	margin: 0;
