+++ frontend/src/styles/variables.scss
@@ -1,0 +1,5 @@
+$text-black: #000;
+$text-bright: #fff;
+$color-background: #1d3461;
+$color-accent: #2589bd;
+$color-highlight: #44a5da;
